core/requirements.txt
fastapi==0.64.0
pydantic==1.8.2
pytest==6.2.4
pytest-asyncio==0.15.1
python-dotenv==0.19.0
pytest-cov==3.0.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.15
alembic==1.6.5
passlib==1.7.4
pydantic[dotenv]
uvicorn==0.13.4
python-jose==3.3.0
python-multipart==0.0.5
httpx==0.18.2
bcrypt==3.2.2
sqlalchemy-utils==0.38.2
requests==2.28.1
dateutils==0.6.12